requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pillow==10.3.0
pydantic==2.7.4
pydantic_core==2.18.4
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
yarl==1.9.4
